requirements.txt
chart_studio==1.1.0
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
streamlit==1.5.1
